Cleanup for tests
